Rename serialize.rs -> codec.rs
