+++ server/src/io/worker.rs
@@ -10,8 +10,8 @@ use crate::config::Config;
 use crate::io::initialhandler::{Action, InitialHandler};
 use crate::io::{ListenerToServerMessage, NewClientInfo, ServerToWorkerMessage};
 use crate::PlayerCount;
+use feather_core::network::codec::MinecraftCodec;
 use feather_core::network::packet::PacketDirection;
-use feather_core::network::serialize::MinecraftCodec;
 use futures::{select, StreamExt};
 use futures::{FutureExt, SinkExt};
 use std::net::SocketAddr;
